Track seconds and use events (specifically, SCREENSHOT_SUCCEEDED)
* Ansel.lua: Two major changes: keep track of seconds, and use events

1. Currently, GetGameTime returns the server time in hours and
minutes.  For our purposes, that's not enough: since screenshot file
names contain the hour, minute, and second when they were taken, and
since we need to know the screenshot file name, we need to know the
second when the screenshot is taken.  For that, I've adoped a hack
that I found on the WowWiki:

  http://www.wowwiki.com/GameTime:Get

This module keeps track of the second.  Its accuracy is, perhaps
surprisingly, limited by the current framerate.  It seems to work well
enough for our purposes, though.

2. We're now tracking the SCREENSHOT_SUCCESSFUL event.

I found to my surprise that the Screenshot() function apparently runs
asynchronously.  The sequence of functions

  UIParent:Hide()
  Screenshot()
  UIParent:Show()

intuitively says: hide the UI (equivalent to hitting alt-z in WoW),
then take a screenshot, then reveal the UI.  The screenehots generated
in this way, however, do contain the UI.  What's apparently happening
is that the Screenshot() function completes its exeuction *before* the
screen is actually captured and the file generated.  Control is passed
to UIParent:Show() before the screen gets captured.  One is tempted to
insert useless idle computation between the Screenshot() and the
UIParent:Show(), but this doesn't feel right.  (And there doesn't
appear to be a sleep or wait or idle function, anyway.)

A solution was suggested on #wowuidev on irc.freenode.net: use events.
Specifically, upon loading Ansel now tracks the SCREENSHOT_SUCCEEDED
event.  AnselTakeSingleShot() now essentially records the time when it
is called, hides the UI if that is requested, and takes a screenshot.
The function Ansel_OnEvent now actually does something: when it
receives the event SCREENSHOT_SUCCEEDED, it reveals the UI (if it was
hidden when the screenshot was taken) and computes the time and other
features of the screenshot.