Use system rather than server seconds
* Ansel.lua: Delete Ansel_GameTime module and use system time instead

I discovered that the screenshot file names that get stored under the
World of Warcraft application directory come from the current system
time, *not* from the current server time.  Thus, it is not necessary
to use the Ansel_GameTime module (which I snarfed from wowwiki); it is
enough to get the current system time.

(It is possible that the screenshot filenames are generated with the
help of the current system time together with a time zone offset, and
that I am saying that it is *only* the current system time that
matters because I'm working in the same time zone as Blizzard's
servers (offset of zero).  If that is the case, then we can adapt
later.)